requirements.txt
# score-inator/requirements.txt
google-adk==1.14.1
requests>=2.31.0
python-dotenv>=1.1.1
pydantic>=2.0.0
typing-extensions
